libnet/tcp: check TCP packet checksum
There seems to be an issue here where large packets (those over the interface MTU?) have an invalid checksum. The code itself is correct as tcpdump shows the same checksum errors as libnet in this regard. There is evidence to show that it is caused by NIC tso (tcp segmentation offload) and by extension gso (general segmentation offload). I have attempted to disable these features however could not solve the problem. For now I'll leave this feature out of the main project branch until it is resolved